refactoring and remove changes to ingestion framework
