+++ crates/sui-bridge-indexer/src/lib.rs
@@ -12,6 +12,7 @@ pub mod metrics;
 pub mod models;
 pub mod postgres_manager;
 pub mod schema;
+pub mod sui_checkpoint_ingestion;
 pub mod sui_transaction_handler;
 pub mod sui_transaction_queries;
 pub mod sui_worker;
